+++ iip-gutenblocks.php
@@ -4,7 +4,7 @@
 * Plugin Name: IIP Gutenblocks
 * Plugin URI: https://github.com/IIP-Design/iip-gutenblocks
 * Description: Plugin providing custom content blocks for the WordPress Gutenberg editor
-* Version: 1.0.0
+* Version: 1.1.1
 * Author: Marek Rewers, U.S. Department of State, IIP Office of Design <designdevops@america.gov>
 * Text Domain: iip-gutenblocks
 * License: GNU General Public License v2.0